+++ src/backend/people/api_urls.py
@@ -3,11 +3,11 @@
 from django.conf import settings
 from django.urls import include, path, re_path
 
+from lasuite.oidc_login.urls import urlpatterns as oidc_urls
+from lasuite.oidc_resource_server.urls import urlpatterns as resource_server_urls
 from rest_framework.routers import DefaultRouter
 
 from core.api.client import viewsets
-from core.authentication.urls import urlpatterns as oidc_urls
-from core.resource_server.urls import urlpatterns as resource_server_urls
 
 from mailbox_oauth2.urls import urlpatterns as mailbox_oauth2_urls